// labs/code-quality/src/exercises/04-shopping-cart/model/cart.model.ts
export interface CartProduct {
  id: number
  name: string
  price: number
  quantity: number
}

export const MIN_PRODUCT_QUANTITY = 1

/* 장바구니 상품 수량을 감소할 수 있는지 판별하는 함수 */
export function canDecreaseQuantity(quantity: number) {
  return quantity > MIN_PRODUCT_QUANTITY
}